+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - General improvements to Message Input View
 - Added ReactionViewClickListener
 - Added support for banning and unbanning users
+- Added support for deleting a channel
+- Add support for switching users via `client.disconnect` and `client.setUser`
+- Add `reload` method to `ChannelListViewModel`
+
+#### Breaking changes:
+
+- Nothing yet
 
 ## September 28, 2019 - 2.1.0
 
@@ -22,7 +29,7 @@
 - Simplify random access to channels
 - Channel query and watch methods now work the same as they do on all other SDKs
 
-### Breaking changes:
+#### Breaking changes:
 
 - `channel.query` does not watch the channel anymore, to retrieve channel state and watch use `channel.watch`
 - `client.getChannelByCID` is now private, use one of the `client.channel` methods to get the same (no null checks needed)